backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]
bcrypt==4.2.0
python-multipart==0.0.17
alembic==1.14.0
aiosqlite==0.20.0
psycopg2-binary>=2.9.9
httpx==0.27.0
python-dotenv==1.0.1
websockets==13.0
jinja2>=3.1.0
itsdangerous>=2.1.0
slowapi==0.1.9
redis==5.2.0
prometheus-client==0.21.0
boto3>=1.35.0
cryptography>=42.0.0
